patel-backend-test-1/requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
strawberry-graphql-django==0.3rc1